requirements.txt
Flask
requests
python-dotenv
GitPython